# backend/specs/1-integrate-rag-chatbot/contracts/openapi.yaml
# OpenAPI Contract: Integrated RAG Chatbot API

openapi: 3.0.0
info:
  title: Integrated RAG Chatbot API
  version: 1.0.0
  description: API for the Integrated RAG Chatbot that allows users to query book content using retrieval augmented generation.
  
servers:
  - url: https://api.rag-chatbot.example.com/v1
    description: Production server
  - url: https://sandbox.rag-chatbot.example.com/v1
    description: Sandbox server for testing

paths:
  /auth/login:
    post:
      summary: Authenticate user and get access token
      description: Authenticates a user and returns a JWT token for subsequent API calls
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "securePassword123"
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type:
                    type: string
                    example: "bearer"
        '401':
          description: Invalid credentials
          
  /books:
    post:
      summary: Upload and process a new book
      description: Uploads and processes a new book for the RAG system
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - title
                - author
              properties:
                file:
                  type: string
                  format: binary
                  description: Book file (PDF, EPUB)
                title:
                  type: string
                  description: Title of the book
                  example: "Introduction to Machine Learning"
                author:
                  type: string
                  description: Author of the book
                  example: "John Doe"
      responses:
        '201':
          description: Book successfully uploaded and processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookContent'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized access
          
  /books/{bookId}:
    get:
      summary: Get book information
      description: Retrieves information about a specific book
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookContent'
        '401':
          description: Unauthorized access
        '404':
          description: Book not found
          
  /query/{bookId}:
    post:
      summary: Query the book content using RAG
      description: Submit a query about the book content and get a response with citations
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: The question or query about the book content
                  example: "What are the main advantages of deep learning?"
                mode:
                  type: string
                  enum: [full_book, selected_text]
                  description: Query mode - full book or selected text only
                  default: full_book
                selected_text:
                  type: string
                  description: Specific text to focus on when mode is 'selected_text'
                  example: "Deep learning is a subset of machine learning that uses neural networks with multiple layers..."
      responses:
        '200':
          description: Query processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Invalid query
        '401':
          description: Unauthorized access
        '404':
          description: Book not found
          
  /embed/{bookId}:
    get:
      summary: Get embeddable chatbot code
      description: Retrieves the HTML/JS code to embed the chatbot in books
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Embeddable code retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  embed_code:
                    type: string
                    description: HTML/JS code to embed the chatbot
                    example: "<div id='rag-chatbot-container'>...</div><script>...</script>"
        '401':
          description: Unauthorized access
        '404':
          description: Book not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    BookContent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the book
        title:
          type: string
          description: The title of the book
        author:
          type: string
          description: The author of the book
        isbn:
          type: string
          description: ISBN identifier for the book
        content_type:
          type: string
          enum: [PDF, EPUB, TEXT, OTHER]
          description: Format of the original content
        processing_status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
          description: Current status of content processing
        created_at:
          type: string
          format: date-time
          description: Timestamp when the book was added
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the book was last updated
          
    QueryResponse:
      type: object
      properties:
        response:
          type: string
          description: The AI-generated response to the query
          example: "Deep learning has several advantages, including its ability to automatically learn hierarchical feature representations..."
        citations:
          type: array
          items:
            type: object
            properties:
              source_id:
                type: string
                description: ID of the source passage
              text:
                type: string
                description: The text of the cited passage
              page_number:
                type: integer
                description: Page number in the original book
          description: List of citations supporting the response
        latency:
          type: number
          format: float
          description: Time taken to process the query in seconds
          example: 1.24
          
    UserQuery:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the query
        book_id:
          type: string
          format: uuid
          description: Reference to the book being queried
        query_text:
          type: string
          description: The original question or prompt submitted
        query_mode:
          type: string
          enum: [FULL_BOOK, SELECTED_TEXT]
          description: Whether querying full book or selected text
        response_text:
          type: string
          description: The AI-generated response
        citation_references:
          type: array
          items:
            type: string
          description: Links to specific source passages
        created_at:
          type: string
          format: date-time
          description: Timestamp when the query was submitted